Extracted dkgExecutor logic out of the node structure
dkgExecutor is similar to the signingExecutor. It is a component responsible for
the full execution of ECDSA Distributed Key Generation: determining members
selected to the signing group, executing off-chain protocol, and publishing the
result to the chain.

The code was extracted out of node with some small changes:
- `joinDKGIfEligible` code has been split into four functions that are hopefully
  easier to follow. `generateSigningGroup` is something that can hopefully be
  split even further but I will approach it in separate PR.
- +1 on the member index is not performed in `checkEligibility` instead of
  before creating the goroutine.
